Week 8-Project (JSON Web Token)

+++ controllers/bookController.ts
@@ -3,13 +3,14 @@ import bookQueries from '../db/queries/bookQueries';
 import Book from '../models/book';
 import bookValidation from '../validators/bookValidator';
 import CustomError from '../errors/CustomError';
-import asyncErrorHandler from '../utils/asyncErrorHandler';
+import asyncHandler from 'express-async-handler';
+import checkPermissions from '../utils/checkPermissions';
 
 class BookController {
   private static bookQueries = new bookQueries();
 
   // Get All Books
-  getAllBooks = asyncErrorHandler(async function (req: Request, res: Response) {
+  getAllBooks = asyncHandler(async function (req: Request, res: Response) {
     const books = await BookController.bookQueries.getAllBooks();
 
     res.status(200).json({
@@ -19,10 +20,10 @@ class BookController {
   });
 
   // Create Book
-  addBook = asyncErrorHandler(async function (req: Request, res: Response) {
+  addBook = asyncHandler(async function (req: Request, res: Response) {
     const bookBody: Book = req.body;
 
-    const { error, value } = bookValidation.validate(bookBody);
+    const { error, value: book } = bookValidation.validate(bookBody);
     if (error) throw new CustomError(400, error.details[0].message);
 
     const existingBook = await BookController.bookQueries.getBookByName(
@@ -31,7 +32,9 @@ class BookController {
 
     if (existingBook) throw new CustomError(400, 'Book already exists');
 
-    await BookController.bookQueries.addBook(value);
+    book.userId = req.userId;
+
+    await BookController.bookQueries.addBook(book);
 
     res.status(201).json({
       status: 201,
@@ -40,7 +43,7 @@ class BookController {
   });
 
   // Get Book By Id
-  getBookById = asyncErrorHandler(async function (req: Request, res: Response) {
+  getBookById = asyncHandler(async function (req: Request, res: Response) {
     const bookId = Number(req.params.id);
 
     const book = await BookController.bookQueries.getBookById(bookId);
@@ -54,15 +57,17 @@ class BookController {
   });
 
   // Update Book
-  updateBook = asyncErrorHandler(async function (req: Request, res: Response) {
+  updateBook = asyncHandler(async function (req: Request, res: Response) {
     const bookId = Number(req.params.id);
     const bookBody = req.body;
 
-    const book = await BookController.bookQueries.getBookById(bookId);
+    const book: any = await BookController.bookQueries.getBookById(bookId);
     if (!book) throw new CustomError(404, 'Book not found');
 
     const { error, value } = bookValidation.validate(bookBody);
     if (error) throw new CustomError(400, error.details[0].message);
+
+    checkPermissions(req.userId, book.userId);
 
     await BookController.bookQueries.updateBook(bookId, value);
 
@@ -73,12 +78,14 @@ class BookController {
   });
 
   // Delete Book
-  deleteBook = asyncErrorHandler(async function (req: Request, res: Response) {
+  deleteBook = asyncHandler(async function (req: Request, res: Response) {
     const bookId = Number(req.params.id);
 
-    const book = await BookController.bookQueries.getBookById(bookId);
+    const book: any = await BookController.bookQueries.getBookById(bookId);
 
     if (!book) throw new CustomError(404, 'Book not found');
+
+    checkPermissions(req.userId, book.userId);
 
     await BookController.bookQueries.deleteBook(bookId);
